refactor:[#13] IntroPage에서 영화 검색 기능을 MoviesSearch 컴포넌트로 이동
